requirements.txt
Django==2.1.3
djangorestframework==3.9.0
django-environ==0.4.5
django-cors-headers==2.4.0
psycopg2-binary=2.7.6.1
zappa-django-utils=0.4.0
